# nextjs/public/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: ESUS Audit AI API
  description: |
    API documentation for the ESUS Audit AI platform.

    This API provides endpoints for managing projects, documents, analysis, and chat functionality.
  version: 1.0.0
  contact:
    name: ESUS Audit AI Support
    email: support@example.com
  license:
    name: Proprietary
servers:
  - url: https://api.example.com/api
    description: Production server
  - url: https://staging.example.com/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Log out a user
      description: Invalidates the current session token
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful logout
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /analysis/document/{id}:
    get:
      tags:
        - Analysis
      summary: Get document analysis results
      description: Returns analysis results for a document
      operationId: getDocumentAnalysis
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Document ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResult'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Analysis not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/users/{id}:
    put:
      tags:
        - Admin
      summary: Update a user
      description: Updates a user by ID (admin only)
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                role:
                  type: string
                  enum: [user, auditor, admin]
                company:
                  type: string
                organization_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User updated successfully
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        auth_user_id:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [user, auditor, admin, super_admin]
        company:
          type: string
        organization_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AnalysisResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        extracted_data:
          type: object
          properties:
            content:
              type: string
            tables:
              type: array
              items:
                type: object
            keyValuePairs:
              type: object
              additionalProperties:
                type: string
        ai_summary:
          type: string
        red_flags:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        processing_time_ms:
          type: integer
        organization_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        type:
          type: string
          enum:
            [
              VALIDATION,
              AUTHENTICATION,
              AUTHORIZATION,
              NOT_FOUND,
              CONFLICT,
              EXTERNAL_SERVICE,
              DATABASE,
              RATE_LIMIT,
              INTERNAL,
            ]
        details:
          type: object
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT